Protocol refactoring fix, update Engine @4.68.0, SDLC @0.179.1
