+++ pom.xml
@@ -100,8 +100,8 @@
 
         <!-- Legend dependency versions -->
         <legend.shared.version>0.25.7</legend.shared.version>
-        <legend.engine.version>4.67.5</legend.engine.version>
-        <legend.sdlc.version>0.178.2</legend.sdlc.version>
+        <legend.engine.version>4.68.0</legend.engine.version>
+        <legend.sdlc.version>0.179.1</legend.sdlc.version>
 
 
         <!-- Dependency versions -->
